Fix リファクタリングを実施した

+++ src/Omnix.Serialization/Base58Btc.cs
@@ -1,8 +1,5 @@
 using System;
 using System.Buffers;
-using System.IO;
-using System.IO.Pipelines;
-using System.Text;
 using Omnix.Base;
 
 namespace Omnix.Serialization
@@ -55,7 +52,7 @@ namespace Omnix.Serialization
             int length = 0;
 
             // Allocate enough space in big-endian base58 representation.
-            int b58Length = (int)(span.Length - zeroCount) * 138 / 100 + 1; // log(256) / log(58), rounded up.
+            int b58Length = (span.Length - zeroCount) * 138 / 100 + 1; // log(256) / log(58), rounded up.
 
             using (var b58 = MemoryPool<byte>.Shared.Rent(b58Length))
             {
@@ -274,7 +271,7 @@ namespace Omnix.Serialization
             int length = 0;
 
             // Allocate enough space in big-endian base256 representation.
-            int b256Length = (int)(text.Length - zeroCount) * 733 / 1000 + 1; // log(58) / log(256), rounded up.
+            int b256Length = (text.Length - zeroCount) * 733 / 1000 + 1; // log(58) / log(256), rounded up.
 
             using (var b256 = MemoryPool<byte>.Shared.Rent(b256Length))
             {
